Show UI elements at the same time
Due to the lazy loading introduced in 573fa4d3118fdccdb00b56d2af724ecc203dcf2b,
the footer controls and search button is visible on the page before the
hamburger menu. This was due to the menu being loaded entirely from the
JS file.

Now the hamburger menu is moved into the HTML, hidden by default and
then displayed when the JS is parsed. The footer is also hidden by
default and displayed when the JS is parsed.